Remove dtype for python 2.7

+++ circletracking/find.py
@@ -23,8 +23,7 @@ def find_disks(image, size_range, number_of_disks=100):
     """ Locate blobs in the image by using a Laplacian of Gaussian method """
     number_of_disks = int(np.round(number_of_disks))
     radii = np.linspace(size_range[0], size_range[1],
-                        num=min(abs(size_range[0] - size_range[1]) * 2.0, 30),
-                        dtype=np.float)
+                        num=min(abs(size_range[0] - size_range[1]) * 2.0, 30))
 
     # Find edges
     edges = canny(image)
